+++ lib/BasicsC/files.c
@@ -794,10 +794,8 @@ bool TRI_ReadPointer (int fd, void* buffer, size_t length) {
 
   ptr = buffer;
 
-
   while (0 < length) {
-    ssize_t n = TRI_READ(fd, ptr, length);
-
+    ssize_t n = TRI_READ(fd, ptr, (unsigned int) length);
 
     if (n < 0) {
       TRI_set_errno(TRI_ERROR_SYS_ERROR);
@@ -827,7 +825,7 @@ bool TRI_WritePointer (int fd, void const* buffer, size_t length) {
   ptr = buffer;
 
   while (0 < length) {
-    ssize_t n = TRI_WRITE(fd, ptr, length);
+    ssize_t n = TRI_WRITE(fd, ptr, (unsigned int) length);
 
     if (n < 0) {
       TRI_set_errno(TRI_ERROR_SYS_ERROR);
@@ -979,7 +977,7 @@ int TRI_CreateLockFile (char const* filename) {
   pid = TRI_CurrentProcessId();
   buf = TRI_StringUInt32(pid);
 
-  rv = TRI_WRITE(fd, buf, strlen(buf));
+  rv = TRI_WRITE(fd, buf, (unsigned int) strlen(buf));
 
   if (rv == -1) {
     res = TRI_set_errno(TRI_ERROR_SYS_ERROR);
